% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DonutMap-package.R
\docType{package}
\name{DonutMap-package}
\alias{DonutMap}
\alias{DonutMap-package}
\title{DonutMap: donut maps with sf and ggplot2}
\description{
DonutMap creates donut charts positioned on a map from tidy data. The main
static entry point is \code{\link[=donut_map]{donut_map()}}. The main interactive entry point is
\code{\link[=donut_leaflet]{donut_leaflet()}}. Use \code{\link[=donut_polygons]{donut_polygons()}} when you need the \code{sf} polygon
layer directly, and \code{\link[=flow_lines]{flow_lines()}} when you need origin-destination line or
trajectory geometries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://aureliennicosiaulaval.github.io/DonutMap/}
  \item \url{https://github.com/AurelienNicosiaULaval/DonutMap}
  \item Report bugs at \url{https://github.com/AurelienNicosiaULaval/DonutMap/issues}
}

}
\author{
\strong{Maintainer}: Aurélien Nicosia \email{116816597+AurelienNicosiaULaval@users.noreply.github.com}

}
\keyword{internal}
